출력문제 추가

+++ cpp/p11382.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    double A, B, C;
+    long long int A, B, C;
     cin >> A >> B >> C;
     cout << A+B+C;
     return 0;
